allow : in SSID
SSIDs can contain the colon character, so we should be robust to that.

See https://www.cisco.com/assets/sol/sb/WAP321_Emulators/WAP321_Emulator_v1.0.0.3/help/Wireless05.html
which says

“The printable characters plus the space (ASCII 0x20) are allowed, but
these six characters are not: ?, ", $, [, \, ], and +.”

This would seem to imply that Unicode is not allowed, and that the
SSID must be printable.

However, the IEEE standard seemed to imply that any string of 32
octets is allowed, including embedded zeros. Later versions of the
IEEE standard may have restricted that to utf8. And there may be a
standardized layer of encoding into an ascii printable character set.
It all seems a bit muddled, honestly, but I suppose they do have to
allow non-latin character sets.