Removed `get_ssid()` from being the default for --ssid, fixes #41
The issues is that `get_ssid()` gets executed even though --ssid has not
been called. This must be because it is the default value for the --ssid
argument.

By setting `get_ssid()` as the default value later in the code prevents
the dependency checking for Network Manager on Linux whenever a user
runs `wifi-password --help` to view the usage.