Made the cli arguments more intuitive and cleaned up the code
There had been confusions regarding the CLI arguments as they could have
been used in the wrong way causing nothing to happen. This has been
fixed by having CLI arguments that are easily understandable for a new
user.

Cleaning was done to the code in order to enhance the legibility. This
was done by fixing the formatting and adding type hinting.